requirements.txt
toml==0.9.2
gunicorn==19.3.0
django-heroku==0.3.1
django-environ==0.4.
Pillow==4.3.0.